Moves PUT endpoint for submitting feedback from submissions.ts to feedback.ts. Updates the endpoint route to remove SubmissionId which was unnecessary/didn't make sense.
