+++ src/api/routes/submissions.ts
@@ -17,7 +17,6 @@ import { Roles } from '../../interfaces/roles.ts';
 import { INewSubmission } from '../../interfaces/submissions.ts';
 import RumbleFeedbackModel from '../../models/rumbleFeedback.ts';
 import SubmissionModel from '../../models/submissions.ts';
-import RumbleService from '../../services/rumble.ts';
 import SubmissionService from '../../services/submission.ts';
 import authHandler from '../middlewares/authHandler.ts';
 import upload from '../middlewares/upload.ts';
@@ -29,7 +28,7 @@ export default (app: IRouter) => {
   const logger: log.Logger = serviceCollection.get('logger');
   const subServiceInstance = serviceCollection.get(SubmissionService);
   const feedbackModelInstance = serviceCollection.get(RumbleFeedbackModel);
-  const rumbleServiceInstance = serviceCollection.get(RumbleService);
+
   app.use(['/submit', '/submission', '/submissions'], route);
 
   // api/submissions/
@@ -221,20 +220,6 @@ export default (app: IRouter) => {
           parseInt(req.params.id, 10),
           parseInt(req.params.flagId, 10)
         );
-        res.setStatus(204).end();
-      } catch (err) {
-        logger.error(err);
-        throw err;
-      }
-    }
-  );
-
-  route.put(
-    '/:submissionId/feedback',
-    authHandler(),
-    async (req: Request, res: Response) => {
-      try {
-        await rumbleServiceInstance.addScoresToFeedback(req.body);
         res.setStatus(204).end();
       } catch (err) {
         logger.error(err);
